anova.hopit {hopit}	R Documentation

Likelihood Ratio Test Tables

Description

Compute likelihood ratio test for two or more hopit objecs.

Usage

## S3 method for class 'hopit'
anova(object, ..., method = c("sequential", "with.most.complex", 'with.least.complex'),
direction = c("decreasing", "increasing"))

Arguments

object
an object containing the results returned by a hopit.

...
additional objects of the same type.

method
the method of model comparison. Choose "sequential" for 1-2, 2-3, 3-4, ... comparisons or "with.most.complex" for 1-2, 1-3, 1-4, ... comparisons, where 1 is the most complex model.

direction
determine if complexity of listed models is "increasing" or "decreasing" (default).

Value

a vector or a matrix with results of the test(s).

Author(s)

Maciej J. Danko

See Also

print.anova.hopit, print.lrt.hopit, lrt.hopit, hopit.

Examples

# DATA
data(healthsurvey)

# the order of response levels is decreasing (from the best health to the worst health)
levels(healthsurvey$health)

## [1] "Excellent" "Very good" "Good"      "Fair"      "Poor"

# Example 1 ---------------------

# fitting two nested models
model1 <- hopit(latent.formula = health ~ hypertension + high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese + diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# model with interaction between hypertension and high_cholesterol
model2 <- hopit(latent.formula = health ~ hypertension * high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese + diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# Likelihood ratio test
lrt1 <- anova(model1, model2)
lrt1

## Full model:
## -- Formula (latent variables): 
## health ~ hypertension + high_cholesterol + heart_attack_or_stroke +  
##     poor_mobility + very_poor_grip + depression + respiratory_problems +  
##     IADL_problems + obese + diabetes + other_diseases + hypertension:high_cholesterol
## -- Formula (threshold variables): ~sex + ageclass + country
## -- Theta: FALSE
## 
## Nested model:
## -- Formula (latent variables): 
## health ~ hypertension + high_cholesterol + heart_attack_or_stroke +  
##     poor_mobility + very_poor_grip + depression + respiratory_problems +  
##     IADL_problems + obese + diabetes + other_diseases
## -- Formula (threshold variables): ~sex + ageclass + country
## -- Theta: FALSE
## 
## Likelihood ratio test:
##  Chi^2 df Pr(>Chi^2)
##   2.55  1       0.11

# print results in shorter form
print(lrt1, short = TRUE)

## 
## Likelihood ratio test:
##  Chi^2 df Pr(>Chi^2)
##   2.55  1       0.11

# equivalently
lrt.hopit(model2, model1)

## Full model:
## -- Formula (latent variables): 
## health ~ hypertension + high_cholesterol + heart_attack_or_stroke +  
##     poor_mobility + very_poor_grip + depression + respiratory_problems +  
##     IADL_problems + obese + diabetes + other_diseases + hypertension:high_cholesterol
## -- Formula (threshold variables): ~sex + ageclass + country
## -- Theta: FALSE
## 
## Nested model:
## -- Formula (latent variables): 
## health ~ hypertension + high_cholesterol + heart_attack_or_stroke +  
##     poor_mobility + very_poor_grip + depression + respiratory_problems +  
##     IADL_problems + obese + diabetes + other_diseases
## -- Formula (threshold variables): ~sex + ageclass + country
## -- Theta: FALSE
## 
## Likelihood ratio test:
##  Chi^2 df Pr(>Chi^2)
##   2.55  1       0.11

# Example 2 ---------------------

# fitting additional nested models
model3 <- hopit(latent.formula = health ~ hypertension * high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese * diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

model4 <- hopit(latent.formula = health ~ hypertension * high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese * diabetes + other_diseases,
              thresh.formula = ~ sex * ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# sequential likelihood ratio tests
# model complexity increases so direction = "increasing"
anova(model1, model2, model3, model4,
      direction = "increasing", method = "sequential")

## Anova (LRTs):
## Method: "sequential"
## 
##                   Chi^2.full df Pr(>Chi^2).full   
## model2 vs. model1      2.550  1         0.11030   
## model3 vs. model2      0.509  1         0.47568   
## model4 vs. model3     26.427 12         0.00934 **
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# likelihood ratio tests of the most complex model with the rest
anova(model1, model2, model3, model4,
      direction = "increasing", method = "with.most.complex")

## Anova (LRTs):
## Method: "with.most.complex"
## 
##                   Chi^2.full df Pr(>Chi^2).full   
## model4 vs. model1     29.485 14         0.00898 **
## model4 vs. model2     26.935 13         0.01270 * 
## model4 vs. model3     26.427 12         0.00934 **
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# likelihood ratio tests of the least complex model with the rest
anova(model1, model2, model3, model4,
      direction = "increasing", method = "with.least.complex")

## Anova (LRTs):
## Method: "with.least.complex"
## 
##                   Chi^2.full df Pr(>Chi^2).full   
## model2 vs. model1      2.550  1         0.11030   
## model3 vs. model1      3.059  2         0.21669   
## model4 vs. model1     29.485 14         0.00898 **
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

[Package hopit version 0.4.167 Index]