boot_hopit_CI {hopit}	R Documentation

Calculating confidence intervals of a boot object using percentile method

Description

Calculating confidence intervals of a boot object using percentile method

Usage

boot_hopit_CI(boot, alpha = 0.05, bounds = c("both", "lo", "up"))

Arguments

boot
a boot object calculated by boot_hopit .

alpha
significance level.

bounds
one of "both", "lo", "up".

Author(s)

Maciej J. Danko

See Also

boot_hopit, getLevels, getCutPoints, latentIndex, standardiseCoef, hopit.

Examples

# see examples in boot_hopit() function.

[Package hopit version 0.4.167 Index]