healthsurvey {hopit}	R Documentation

Artificially generated health survey data

Description

A dataset containing artificially generated survey data

Usage

healthsurvey

Format

A data frame with 10000 rows and 11 variables:

ID

personal identification number.

health

reported health, 5 levels.

diabetes

has diabetes? "yes" or "no"?

obese

has obese? "yes" or "no"?

IADL_problems

problems in Instrumental Activities of Daily Living? "yes" or "no"?

hypertension

has hypertension? "yes" or "no"?

high_cholesterol

has high cholesterol? "yes" or "no"?

respiratory_problems

has respirator problems? "yes" or "no"?

heart_attack_or_stroke

had stroke or heart attack? "yes" or "no"?

poor_mobility

has poor mobility? "yes" or "no"?

very_poor_grip

cannot perform grip strength? "yes" or "no"?

depression

has depression? "yes" or "no"?

other_diseases

has other diseases? "yes" or "no"?

sex

sex/gender: woman or man.

ageclass

categorized age: [50,60), [60,70), [70,80), [80,120).

education

two levels of education: primary or lower and secondary or higher.

country

country: X, Y, or Z.

csw

cross-sectional survey weights.

psu

primary statistical unit.

ssu

secondary statistical unit.

Source

Data was randomly generated using probabilities of occurrence of particular combination of diseases, conditions, sex, age, education, reported health, etc. The structure of the data and some probabilities were inspired by WAVE1 SHARE database (DOIs: 10.6103/SHARE.w1.600), see Börsch-Supan et al for methodological details (Börsch-Supan et al. 2013).

The SHARE data collection has been primarily funded by the European Commission through FP5 (QLK6-CT-2001-00360), FP6 (SHARE-I3: RII-CT-2006-062193, COMPARE: CIT5-CT-2005-028857, SHARELIFE: CIT4-CT-2006-028812) and FP7 (SHARE-PREP: N°211909, SHARE-LEAP: N°227822, SHARE M4: N°261982). Additional funding from the German Ministry of Education and Research, the Max Planck Society for the Advancement of Science, the U.S. National Institute on Aging (U01_AG09740-13S2, P01_AG005842, P01_AG08291, P30_AG12815, R21_AG025169, Y1-AG-4553-01, IAG_BSR06-11, OGHA_04-064, HHSN271201300071C) and from various national funding sources is gratefully acknowledged (see www.share-project.org).

None of the individuals (records) represent any true individual (record) of the SHARE database

References

Börsch-Supan A, Brandt M, Hunkler C, et al (2013) Data resource profile: The survey of health, ageing and retirement in europe (share). Int J Epidemiol 42:992–1001. doi: 10.1093/ije/dyt088

Examples

# load *healthsurvey* dataset
data(healthsurvey)

# horizontal view on the dataset (omitting ID)
print(t(healthsurvey[1:6,-1]), quote=FALSE, na.print='NA', right=TRUE)

##                                1       2       3       4         5       6
## health                 Very good    Good    Good    Good Excellent    Good
## diabetes                      no      no     yes      no        no      no
## obese                         no      no      no      no        no      no
## IADL_problems                 no      no      no      no        no      no
## high_cholesterol              no     yes      no      no        no     yes
## respiratory_problems          no      no      no      no        no     yes
## poor_mobility                 no      no      no      no        no     yes
## very_poor_grip                no      no      no      no        no      no
## depression                    no      no      no     yes        no      no
## other_diseases               yes     yes      no      no        no     yes
## sex                          man     man     man     man     woman     man
## ageclass                     80+   70-79   50-59   60-69       80+     80+
## education                  prim-   prim-   prim-    sec+     prim-   prim-
## country                        Y       Y       X       Y         Z       Y
## csw                      2407.48 1198.12  885.26  772.04   1304.24  917.16
## psu                           YB      YB      XC      YA        ZB      YD
## ssu                           NA      NA   XCgis      NA        NA      NA
## hypertension                  no     yes      no      no        no     yes
## heart_attack_or_stroke        no     yes      no      no        no      no

[Package hopit version 0.4.167 Index]