hopit.control {hopit}	R Documentation

Auxiliary for controlling the fitting of hopit model

Description

Auxiliary function for controlling the fitting of hopit model. Use this function to set control parameters of the hopit and other related functions.

Usage

hopit.control(grad.eps = 3e-05, bgfs.maxit = 10000, cg.maxit = 10000,
  nlm.maxit = 150, bgfs.reltol = 5e-10, cg.reltol = 5e-10,
  nlm.gradtol = 1e-07, nlm.steptol = 1e-07, fit.methods = c("CG",
  "BFGS"), quick.fit = TRUE, trace = TRUE, thresh.start = -Inf,
  thresh.1.exp = FALSE, LL_out_val = -Inf)

Arguments

grad.eps
epsilon for numerical Hessian function.

bgfs.maxit, cg.maxit, nlm.maxit
the maximum number of iterations. See optim and nlm for details.

bgfs.reltol, cg.reltol
relative convergence tolerance. See optim for details.

nlm.gradtol, nlm.steptol
tolerance at which the scaled gradient is considered close enough to zero and minimum allowable relative step length. See nlm.

fit.methods
either 'CG' or 'BFGS'. See optim.

quick.fit
logical, if TRUE extensive nlm optimization method is ignored and only BFGS and CG methods are run.

trace
logical, if to trace model fitting.

LL_out_val, thresh.1.exp, thresh.start
internal parameters under development, do not change.

Author(s)

Maciej J. Danko

See Also

hopit

[Package hopit version 0.4.167 Index]