summary.hopit {hopit}	R Documentation

Calculate model summary

Description

Calculate model summary

Usage

## S3 method for class 'hopit'
summary(object, robust.se = TRUE, ...)

Arguments

object
an hopit object.

robust.se
logical indicating if to use robust standard errors based on the sandwich estimator. If survey deign is detected than this option is ignored.

...
further arguments passed to or from other methods.

control
a list with control parameters. See hopit.control.

Author(s)

Maciej J. Danko

[Package hopit version 0.4.167 Index]