update.latent {hopit}	R Documentation

INTERNAL : updating model uses new set of latent variable coefficients

Description

INTERNAL : updating model uses new set of latent variable coefficients

Usage

update.latent(model, newregcoef, data)

Arguments

model
a fitted hopit model.

newregcoef
a new set of latent variable coefficients. Vector of length of first element of model$parcount.

data
used to fit original model.

Details

LL gradient and Hessian and variance-covariance matrices are not re-calcualted in this function!

Author(s)

Maciej J. Danko

[Package hopit version 0.4.167 Index]